step2  수정및  파일  분행
