+++ CodeStarterCamp_Week4/Models/BodyCondition.swift
@@ -9,24 +9,15 @@ import Foundation
 
 final class BodyCondition {
     
-    var upperBodyStrength = 0
-    var lowerBodyStrength = 0
-    var muscularEndurance = 0
-    var fatigue = 0
+    var upperBodyStrength: Int
+    var lowerBodyStrength: Int
+    var muscularEndurance: Int
+    var fatigue: Int
     
     init(upperBodyStrength: Int, lowerBodyStrength: Int,  muscularEndurance: Int, fatigue: Int) {
         self.upperBodyStrength = upperBodyStrength
         self.lowerBodyStrength = lowerBodyStrength
         self.muscularEndurance = muscularEndurance
         self.fatigue = fatigue
-    }
-    
-    func printBodyCondition() {
-        print("--------------")
-        print("현재의 컨디션은 다음과 같습니다 ")
-        print("상체근력 : \(upperBodyStrength)")
-        print("하체근력: \(lowerBodyStrength)")
-        print("근지구력: \(muscularEndurance)")
-        print("피로도: \(fatigue)")
     }
 }
