+++ CodeStarterCamp_Week4/Views/ResultView.swift
@@ -1,0 +1,8 @@
+//
+//  ResultView.swift
+//  CodeStarterCamp_Week4
+//
+//  Created by 서원지 on 2022/01/05.
+//
+
+import Foundation
